+++ src/Keyvault/Set-Keyvault-Permissions-for-AppService-Identity.ps1
@@ -7,7 +7,8 @@ param (
     [Parameter()][string] $KeyvaultSecretPermissions,
     [Parameter()][string] $KeyvaultStoragePermissions,
     [Parameter(Mandatory)][string] $KeyvaultName,
-    [Parameter()][string] $AppServiceSlotName
+    [Parameter()][string] $AppServiceSlotName,
+    [Parameter()][bool] $ApplyToAllSlots = $false
 )
 
 #region ===BEGIN IMPORTS===
@@ -16,13 +17,17 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
-$identityId = Get-ManagedIdentity -AppService -ResourceName $AppServiceName -ResourceGroupName $AppServiceResourceGroupName -AppServiceSlotName $AppServiceSlotName
+if ($ApplyToAllSlots)
+{
+    $slots = (Invoke-Executable az webapp deployment slot list --resource-group $AppServiceResourceGroupName --name $AppServiceName | ConvertFrom-Json)
+    foreach ($slot in $slots)
+    {
+        $identityId = Get-ManagedIdentity -AppService -ResourceName $AppServiceName -ResourceGroupName $AppServiceResourceGroupName -AppServiceSlotName $slot.name
+        Set-KeyvaultPermissions -KeyvaultName $KeyvaultName -ManagedIdentityId $identityId -KeyvaultCertificatePermissions $KeyvaultCertificatePermissions -KeyvaultKeyPermissions $KeyvaultKeyPermissions -KeyvaultSecretPermissions $KeyvaultSecretPermissions -KeyvaultStoragePermissions $KeyvaultStoragePermissions
+    }
+}
 
-$kvcp = $KeyvaultCertificatePermissions -split ' '
-$kvkp = $KeyvaultKeyPermissions -split ' '
-$kvsp = $KeyvaultSecretPermissions -split ' '
-$kvstp = $KeyvaultStoragePermissions -split ' '
-
-Invoke-Executable az keyvault set-policy --certificate-permissions @kvcp --key-permissions @kvkp --secret-permissions @kvsp --storage-permissions @kvstp --object-id $identityId --name $KeyvaultName
+$identityId = Get-ManagedIdentity -AppService -ResourceName $AppServiceName -ResourceGroupName $AppServiceResourceGroupName -AppServiceSlotName:$AppServiceSlotName
+Set-KeyvaultPermissions -KeyvaultName $KeyvaultName -ManagedIdentityId $identityId -KeyvaultCertificatePermissions $KeyvaultCertificatePermissions -KeyvaultKeyPermissions $KeyvaultKeyPermissions -KeyvaultSecretPermissions $KeyvaultSecretPermissions -KeyvaultStoragePermissions $KeyvaultStoragePermissions
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet