+++ src/Storage-Accounts/Remove-Public-Access-from-StorageAccount.ps1
@@ -1,0 +1,15 @@
+[CmdletBinding()]
+param (
+    [Parameter(Mandatory)][string] $StorageAccountResourceGroupName,
+    [Parameter(Mandatory)][string] $StorageAccountName
+)
+
+#region ===BEGIN IMPORTS===
+Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
+#endregion ===END IMPORTS===
+
+Write-Header -ScopedPSCmdlet $PSCmdlet
+
+Invoke-Executable az storage account update --resource-group $StorageAccountResourceGroupName --name $StorageAccountName --default-action 'Deny'
+
+Write-Footer -ScopedPSCmdlet $PSCmdlet