+++ src/Users-and-Accounts/Get-ObjectID-for-ServiceUser.ps1
@@ -1,7 +1,8 @@
 [CmdletBinding()]
 param (
     [Parameter(Mandatory)][string] $ServiceUserEmail,
-    [Parameter(Mandatory)][string] $ServiceUserPassword
+    [Parameter(Mandatory)][string] $ServiceUserPassword,
+    [Parameter()][string] $OutputPipelineVariableName = "ServiceUserObjectId"
 )
 
 #region ===BEGIN IMPORTS===
@@ -11,6 +12,10 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
 Invoke-Executable az login --username $ServiceUserEmail --password $ServiceUserPassword --allow-no-subscriptions
-Write-Output (Invoke-Executable az ad user show --id $ServiceUserEmail | ConvertFrom-Json).objectId
+$objectId = (Invoke-Executable az ad user show --id $ServiceUserEmail | ConvertFrom-Json).objectId
+
+Write-Host "##vso[task.setvariable variable=$($OutputPipelineVariableName);isOutput=true]$objectId"
+
+Write-Output $objectId
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet